Base setup for graph optimization

+++ common/src/vkcnn/common/hypergraph/EdgeId.hpp
@@ -2,19 +2,20 @@
 
 #include <cstddef>
 #include <compare>
+#include <cstdint>
 
 namespace vkcnn::hypergraph {
 
 struct EdgeId {
 public:
-  explicit constexpr EdgeId(std::size_t id) : m_id(id) {}
+  explicit constexpr EdgeId(std::uint64_t id) : m_id(id) {}
 
-  constexpr operator std::size_t() const { return m_id; }
+  constexpr operator std::uint64_t() const { return m_id; }
 
   constexpr auto operator<=>(const EdgeId &) const = default;
 
 private:
-  std::size_t m_id;
+  std::uint64_t m_id;
 };
 
 } // namespace vkcnn::hypergraph
